+++ app/commands/index.js
@@ -10,6 +10,7 @@ import COMMAND_BOT_CONTINUE from './bot-continue.js';
 import COMMAND_BOT_DEACTIVATE from './bot-deactivate.js';
 import COMMAND_BOT_DRAW_DEMO from './bot-draw-demo.js';
 import COMMAND_BOT_DRAW from './bot-draw.js';
+import COMMAND_BOT_RETRY from './bot-retry.js';
 import COMMAND_BOT_SUMMON_DEMO from './bot-summon-demo.js';
 import COMMAND_BOT_TALK_DEMO from './bot-talk-demo.js';
 import COMMAND_BOT_TALK from './bot-talk.js';
@@ -42,6 +43,7 @@ export const ALL_COMMANDS = [
   COMMAND_BOT_DEACTIVATE,
   COMMAND_BOT_DRAW_DEMO,
   COMMAND_BOT_DRAW,
+  COMMAND_BOT_RETRY,
   COMMAND_BOT_SUMMON_DEMO,
   COMMAND_BOT_TALK_DEMO,
   COMMAND_BOT_TALK,
@@ -96,6 +98,7 @@ export {
   COMMAND_BOT_DEACTIVATE,
   COMMAND_BOT_DRAW_DEMO,
   COMMAND_BOT_DRAW,
+  COMMAND_BOT_RETRY,
   COMMAND_BOT_SUMMON_DEMO,
   COMMAND_BOT_TALK_DEMO,
   COMMAND_BOT_TALK,