+++ app/handlers/index.js
@@ -7,6 +7,7 @@ import docHandler from './doc.js';
 import drawHandler from './draw.js';
 import enquireHandler from './enquire.js';
 import reportHandler from './report.js';
+import retryHandler from './retry.js';
 import talkHandler from './talk.js';
 import versionHandler from './version.js';
 
@@ -20,6 +21,7 @@ export {
   drawHandler,
   enquireHandler,
   reportHandler,
+  retryHandler,
   talkHandler,
   versionHandler,
 };