+++ app/handlers/retry.js
@@ -1,5 +1,7 @@
+import config from '../../config/index.js';
+import { PARTICIPANT_AI } from '../../services/openai.js';
 import { generateCompletion } from '../../utils/index.js';
-import { ALL_COMMANDS, COMMAND_BOT_CONTINUE } from '../commands/index.js';
+import { COMMAND_BOT_CONTINUE, COMMAND_BOT_RETRY, GENERAL_COMMANDS } from '../commands/index.js';
 import Context from '../context.js';
 import { updateHistory } from '../history/index.js';
 import { getPrompt, setPrompt } from '../prompt/index.js';
@@ -8,7 +10,7 @@ import { getPrompt, setPrompt } from '../prompt/index.js';
  * @param {Context} context
  * @returns {boolean}
  */
-const check = (context) => context.isCommand(COMMAND_BOT_CONTINUE);
+const check = (context) => context.isCommand(COMMAND_BOT_RETRY);
 
 /**
  * @param {Context} context
@@ -18,16 +20,13 @@ const exec = (context) => check(context) && (
   async () => {
     updateHistory(context.id, (history) => history.erase());
     const prompt = getPrompt(context.userId);
-    const { lastSentence } = prompt;
-    if (lastSentence.isEnquiring) prompt.erase();
+    prompt.erase().write(PARTICIPANT_AI);
     try {
       const { text, isFinishReasonStop } = await generateCompletion({ prompt: prompt.toString() });
       prompt.patch(text);
-      if (lastSentence.isEnquiring && !isFinishReasonStop) prompt.write('', lastSentence.text);
       setPrompt(context.userId, prompt);
-      if (!lastSentence.isEnquiring) updateHistory(context.id, (history) => history.patch(text));
-      const defaultActions = ALL_COMMANDS.filter(({ type }) => type === lastSentence.text);
-      const actions = isFinishReasonStop ? defaultActions : [COMMAND_BOT_CONTINUE];
+      updateHistory(context.id, (history) => history.write(config.BOT_NAME, text));
+      const actions = isFinishReasonStop ? [] : [COMMAND_BOT_CONTINUE];
       context.pushText(text, actions);
     } catch (err) {
       context.pushError(err);